+++ include/xtables.h
@@ -218,7 +218,7 @@ enum exittype {
 };
 
 /* this is a special 64bit data type that is 8-byte aligned */
-#define aligned_u64 unsigned long long __attribute__((aligned(8)))
+#define aligned_u64 u_int64_t __attribute__((aligned(8)))
 
 extern void exit_printhelp() __attribute__((noreturn));
 extern void exit_tryhelp(int) __attribute__((noreturn));
